factor lookup out
